Functions to simplify the control of the arrow
I've added 2 functions that allow the user to control the position of the mouse arrow in a very easy and intuitive manner. The first is designed to move the arrow by the numbers x and y and the second uses the first one to move it to a specific x,y on the screen, in this case I needed to tell it to first move it by -10000 so it goes to the very edge of the screen, regardless of the screen size(I didn't know how else to do that). I hope you like it.