fix(sqlite): improve floating-point precision
This adds support for values of type double when binding values to a
query. Prior to this, doubles would be converted to a string and only
be stored with 6 digits of precision in the database.

Additionally, floats stored in the database were converted to strings
when retrieved and therefore be reduced to 6 digits of precision, too.