fix nanovega tainting the active VAO
this is both needed for modern opengl because if you are unlucky with your graphics drivers there it won't render anything when the core context is enabled, but also when rendering custom data and having stray VAOs open that you didn't unbind on the start of the nanovega render calls.

stuff like `glEnableVertexAttribArray` and `glVertexAttribPointer`, which is called below, is only valid with a VAO bound.