Prevent infinite loop on interrupted websocket
If a websocket message is split across continuation frames, it is
possible to get to this inner loop. If the browser is closed in the
middle of the message, the connection is closed, which makes the select
check return true, but then the other check returns false, indicating
eof, but the loop would continue.

We do not close the connection here, it just throws, but that's
consistent with other things for now. It should probably change too.